Mrj 31 les promotions soient consultables par les responsable des rayon seulement entre enterval

+++ src/main/java/ma/yc/api/repository/AdminCentreAuthetificationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminCentreAuthetificationRepository extends JpaRepository<AdminCentre, Integer> {
+public interface AdminCentreAuthetificationRepository extends JpaRepository<AdminCentre, Long> {
     public AdminCentre findByEmail(String email);
 }
